Handle `ERR-INVALID-AUTH` responses from Conduit in `patch` workflow
Conduit responds to requests with either `ERR-INVALID-SESSION` or
`ERR-INVALID-AUTH` if the request wasn't sufficiently authenticated.
Arcanist's `patch` workflow can automatically attempt to recover from
situations in which Conduit responds to unauthenticated requests with
`ERR-INVALID-SESSION` (by resending an authenticated version of the
request), but not `ERR-INVALID-AUTH` - recover from `ERR-INVALID-AUTH`
in the same way.