Allow https.cabundle to be a directory.
Arcanist really, really wants to set a CA bundle.  This is wrong.  It
should rely on Curl to pick up the system trust store and only try to
force its own if the user explicitly requested it.  But for now,
allowing https.cabundle to refer to a directory (which will be passed
to CURL_CAPATH instead of CURL_CAINFO) will make life easier for
integrators on platforms where the system trust store is a directory.